Remove race condition between GhostDriver starting a browser init
Before we had a delay in which it was expected that Ghostdriver
would have started in. This didn't happen on slower machines.

I have replaced it with a class that searches the output of PhantomJS
for a string indicating that GhostDriver is ready to receive
connections, and then kicking off the browser init.